Breaking out akka-http runtime tests
