+++ modules/scala-akka-http/src/test/scala/core/issues/Issue43.scala
