+++ modules/scala-akka-http/src/test/scala/tests/generators/akkaHttp/StaticParametersTest.scala
